fix_api(be): Fixed all use cases related to cost and consumption and added migration to change that fields from decimal to double

+++ backend/ElectroManage.Application/Features/Register/Command/Post/CreateRegisterResponse.cs
@@ -3,8 +3,8 @@
 public record CreateRegisterResponse
 {
     public long Id { get; set; }
-    public decimal Cost { get; set; }
-    public decimal Consumption { get; set; }
+    public double Cost { get; set; }
+    public double Consumption { get; set; }
     public string CompanyName { get; set; } = string.Empty;
     public DateTime Date { get; set; }
 }
