+++ backend/ElectroManage.Application/Mocks/ListMonthlastThreeYearsBogusConfig.cs
@@ -6,7 +6,7 @@ public class ListMonthlastThreeYearsBogusConfig : AutoFaker<YearCostDTO>
 {
     public ListMonthlastThreeYearsBogusConfig()
     {
-        RuleFor(x => x.MeanConsumption, f => Math.Round(f.Random.Decimal(300, 1000),1));
-        RuleFor(x => x.MeanCost, f => Math.Round(f.Random.Decimal(500, 100),2));
+        RuleFor(x => x.MeanConsumption, f => Math.Round(f.Random.Double(300, 1000),1));
+        RuleFor(x => x.MeanCost, f => Math.Round(f.Random.Double(500, 100),2));
     }
 }
